requirements.txt
flask==2.3.3
python-dotenv==1.0.0
langchain==0.0.267
langchain-google-genai==0.0.5
python-telegram-bot==20.4
